feat: 发布1.0.0版本

+++ packages/oumi-swagger-api/lib/utils/fs.js
@@ -4,6 +4,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.writeFile = exports.writeJSON = exports.createFileSync = void 0;
+const path_1 = __importDefault(require("path"));
 const fs_extra_1 = __importDefault(require("fs-extra"));
 const index_1 = require("./index");
 /**
@@ -11,13 +12,14 @@ const index_1 = require("./index");
  * @param filePath
  * @param callback
  */
-exports.createFileSync = (filePath, callback) => {
+const createFileSync = (filePath, callback) => {
     fs_extra_1.default.ensureFile(filePath, (err, data) => {
         if (typeof callback === 'function') {
             callback(err, data);
         }
     });
 };
+exports.createFileSync = createFileSync;
 function writeJSON(filePath, data, callback) {
     fs_extra_1.default.createFileSync(filePath);
     fs_extra_1.default.writeFile(filePath, JSON.stringify(data, null, '\t'), null, (err, data2) => {
@@ -33,9 +35,30 @@ function writeJSON(filePath, data, callback) {
     });
 }
 exports.writeJSON = writeJSON;
-function writeFile(filePath, content) {
+function writeFile(filePath, content, options) {
+    const { allowRepeat = true } = options || {};
     fs_extra_1.default.createFileSync(filePath);
-    fs_extra_1.default.writeFile(filePath, content, null, () => { });
-    index_1.log.green(`writeFile: ${filePath}`);
+    // 不允许文件重复
+    if (!allowRepeat && fs_extra_1.default.existsSync(filePath)) {
+        // eslint-disable-next-line no-inner-declarations
+        function testFile(filePath1, count) {
+            const basename = path_1.default.basename(filePath1);
+            const [name, ext] = basename.split('.');
+            const newname = `${name}_${count}.${ext}`;
+            const newPath = filePath1.replace(basename, newname);
+            if (!fs_extra_1.default.existsSync(newPath)) {
+                fs_extra_1.default.writeFile(newPath, content, null, () => { });
+                index_1.log.green(`writeFile: ${newPath}`);
+            }
+            else {
+                testFile(filePath1, count + 1);
+            }
+        }
+        testFile(filePath, 1);
+    }
+    else {
+        fs_extra_1.default.writeFile(filePath, content, null, () => { });
+        index_1.log.green(`writeFile: ${filePath}`);
+    }
 }
 exports.writeFile = writeFile;
